split graphics cmnd
