+++ modules/graphics/commands/helpers/index.ts
@@ -1,1 +1,1 @@
-export * from "./exportBoard";
+export * from "./makeGameGraphics";
