+++ modules/graphics/package.json
@@ -1,8 +1,9 @@
 {
   "scripts": {
-    "robot": "npm run exportIcons; npm run exportMarks; npm run exportGameGraphics",
+    "robot": "npm run exportIcons; npm run exportMarks; npm run makeGameGraphics; npm run exportGameGraphics",
     "clean": "rm -rf dist; mkdir dist;",
     "exportGameGraphics": "cd ../..; npx babel-node --extensions '.ts','.js' ./modules/graphics/commands/exportGameGraphics",
+    "makeGameGraphics": "cd ../..; npx babel-node --extensions '.ts','.js' ./modules/graphics/commands/makeGameGraphics",
     "exportIcons": "cd ../..; npx babel-node --extensions '.ts','.js' ./modules/graphics/commands/exportIcons",
     "exportMarks": "cd ../..; npx babel-node --extensions '.ts','.js' ./modules/graphics/commands/exportMarks"
   },
